\name{play}
\alias{play}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The Oregon Trail game
}
\description{
Runs the 1978 BASIC Oregon Trail game using a very limited interpreter written in R
}
\usage{
play(debug = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{debug}{
If \code{TRUE}, gratuitous levels of debugging information are printed.
}
}
\details{
THIS PROGRAM SIMULATES A TRIP OVER THE OREGON TRAIL FROM
INDEPENDENCE, MISSOURI TO OREGON CITY, OREGON IN 1847.
YOUR FAMILY OF FIVE WILL COVER THE 2040 MILE OREGON TRAIL
IN 5-6 MONTHS --- IF YOU MAKE IT ALIVE.

YOU HAD SAVED $900 TO SPEND FOR THE TRIP, AND YOU'VE JUST
   PAID $200 FOR A WAGON.
YOU WILL NEED TO SPEND THE REST OF YOUR MONEY ON THE
   FOLLOWING ITEMS:

     OXEN - YOU CAN SPEND $200-$300 ON YOUR TEAM
            THE MORE YOU SPEND, THE FASTER YOU'LL GO
               BECAUSE YOU'LL HAVE BETTER ANIMALS

     FOOD - THE MORE YOU HAVE, THE LESS CHANCE THERE
               IS OF GETTING SICK

     AMMUNITION - 81 BUYS A BELT OF 50 BULLETS
            YOU WILL NEED BULLETS FOR ATTACKS BY ANIMALS
               AND BANDITS, AND FOR HUNTING FOOD

     CLOTHING - THIS IS ESPECIALLY IMPORTANT FOR THE COLD
               WEATHER YOU WILL ENCOUNTER WHEN CROSSING
               THE MOUNTAINS

     MISCELLANEOUS SUPPLIES - THIS INCLUDES MEDICINE AND
               OTHER THINGS YOU WILL NEED FOR SICKNESS
               AND EMERGENCY REPAIRS


YOU CAN SPEND ALL YOUR MONEY BEFORE YOU START YOUR TRIP -
OR YOU CAN SAVE SOME OF YOUR CASH TO SPEND AT FORTS ALONG
THE WAY WHEN YOU RUN LOW.  HOWEVER, ITEMS COST MORE AT
THE FORTS.  YOU CAN ALSO GO HUNTING ALONG THE WAY TO GET
MORE FOOD.
WHENEVER YOU HAVE TO USE YOUR TRUSTY RIFLE ALONG THE WAY,
YOU WILL BE TOLD TO TYPE IN A WORD (ONE THAT SOUNDS LIKE A
GUN SHOT).  THE FASTER YOU TYPE IN THAT WORD AND HIT THE
'RETURN' KEY, THE BETTER LUCK YOU'LL HAVE WITH YOUR GUN.

AT EACH TURN, ALL ITEMS ARE SHOWN IN DOLLAR AMOUNTS
EXCEPT BULLETS
WHEN ASKED TO ENTER MONEY AMOUNTS, DON'T USE A '$'.

GOOD LUCK!!!

}
\value{
None
}
\references{
\url{https://github.com/fortran-gaming/oregon-trail-1975}
}
\author{
Thomas Lumley
}



%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
\dontrun{play()}
}
